Extract EV search helpers for reuse with optimizer work
